Get name resolution working with haystack (some glitches)
popit-resolver previously iterated every person/membership and
performed various string comparisons, which was slow and hard
to debug.  The current work moves the process to haystack
(defaulting to elasticsearch backend) which is faster, and with
a bit of luck should be more flexible and accurate.

We save various "EntityName" records for each person, each being
a variant of the member's name or role.  Currently the constructed
options are:

    - full name
    - name with initials
    - if the person belongs to a party:
        - full name + (party name)
        - name with initials + (party name)
    - for every membership
        - role label + organization name

Without fine-tuning, this approach catches a lot more names, with
a greater risk of false-positives.

Problems:

    - false positives. e.g. the wrong "matlala" / "mmatlala"
    - similarly doesn't seem to penalise extra words so, e.g. "Leader of the Opposition" may match "Deputy Leader of the Opposition"
    - this may be a scoring issue?  debug output suggests that haystack.SearchResult.score = 0 for every record
    - filtering by date filters away every result (presumably an error on my part, have disabled for now)