Try more combinations of initials and honorifics
Previously, only one set of initials would be tried - either those from
the initials field, or the first letters of the given names, separated
by space.  This commit changes this behaviour so that both would be
tried (if they both exist and are distinct) as well as just a single
first initial, and different types of representation of the initials
(space separated and directly abutting).

In addition, these combinations are tried both with and without any
honorifics.

The risk here is that we match lots of names that shouldn't be matched,
but at the moment lots of obvious things are being missed (e.g. 'J Malema'
doesn't match 'Julius Sello Malema' at the moment).

This also checks the 'given_name' field (the one in the Popolo spec) if
there is no 'given_names' field.